pyproject.toml
[tool.poetry]
name = "blurring_as_a_service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.9"
pandas = ">=1.1.4"
GitPython = {version = "^3.1.30", source = "PyPI"}
ipython = {version = "^8.8.0", source = "PyPI"}
matplotlib = ">=3.2.2"
numpy = ">=1.18.5"
opencv-python = ">=4.1.1"
Pillow = ">=7.1.2"
psutil = {version = "^5.9.4", source = "PyPI"}
PyYAML = {version = ">=5.3.1", source = "PyPI"}
requests = {version = ">=2.23.0", source = "PyPI"}
scipy = {version = ">=1.4.1", source = "PyPI"}
thop = {version = ">=0.1.1", source = "PyPI"}
torch = {version = "1.8.0", source = "PyPI"}
torchvision = {version = "0.9.0", source = "PyPI"}
tqdm = {version = ">=4.64.0", source = "PyPI"}
tensorboard = {version = ">=2.4.1", source = "PyPI"}
seaborn = {version = ">=0.11.0", source = "PyPI"}
pre-commit = {version = "^2.21.0", source = "PyPI"}
pytest = {version = "^7.2.1", source = "PyPI"}
azure-ai-ml = {version = "^1.2.0", source = "PyPI"}
mldesigner = {version = "^0.1.0-beta.9", source = "PyPI"}
dataclass-wizard = {version = "^0.22.2", source = "PyPI"}
python-geohash = {version = "^0.8.5", source = "PyPI"}
azure-identity = {version = "^1.12.0", source = "PyPI"}
panorama = {git = "https://github.com/Computer-Vision-Team-Amsterdam/panorama.git", rev = "v0.2.2"}
pycocotools = {version = "^2.0.6", source="PyPI"}
cryptography = "<39.0.0"
markdown-it-py = "^2.2.0"
joblib = "^1.2.0"
azureml-core = "^1.49.0"
setuptools = "^67.6.0"

[tool.poetry.dev-dependencies]

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
